Refactor de la fsm (la recette coffee marche po mais le pb est identifié je réfléchi un peu)

+++ DrinkingFactory/fr.univcotedazur.polytech.si4.fsm.project/src/fr/univcotedazur/polytech/si4/fsm/project/products/Product.java
@@ -2,9 +2,9 @@ package fr.univcotedazur.polytech.si4.fsm.project.products;
 
 public abstract class Product {
     public String name;
-    public int price;
+    public long price;
 
-    Product(String name, int price) {
+    Product(String name, long price) {
         this.name = name;
         this.price = price;
     }
